refactor project tools workspace inventory modules
