+++ .sampo/changesets/835-split-workspace-inventory.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/project-tools: patch
+---
+
+Split workspace inventory types, templates, parser, read helpers, and mutation helpers into focused runtime modules while preserving the public barrel.
